+++ common/src/main/java/net/xstopho/resource_cracker/item/ScytheItem.java
@@ -9,6 +9,7 @@ import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.Attribute;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.SwordItem;
@@ -18,14 +19,13 @@ import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.CropBlock;
 import net.minecraft.world.level.block.state.BlockState;
-import net.xstopho.resource_cracker.registries.AttributeRegistry;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 import java.util.UUID;
 
 public class ScytheItem extends SwordItem {
-    public static final UUID BASE_ATTACK_RANGE_UUID = UUID.nameUUIDFromBytes("CRACKER_BASE_ATTACK_RANGE_UUID".getBytes());
+    public static final UUID BASE_ATTACK_KNOCKBACK_UUID = UUID.nameUUIDFromBytes("BASE_ATTACK_KNOCKBACK_UUID".getBytes());
     private static final int radius = 1;
 
     public ScytheItem(Tier tier, int attackDamage, float attackSpeed) {
@@ -65,7 +65,7 @@ public class ScytheItem extends SwordItem {
     public Multimap<Attribute, AttributeModifier> getDefaultAttributeModifiers(EquipmentSlot slot) {
         ImmutableMultimap.Builder<Attribute, AttributeModifier> attributes = ImmutableMultimap.builder();
         attributes.putAll(super.getDefaultAttributeModifiers(slot));
-        attributes.put(AttributeRegistry.ATTACK_RANGE.get(), new AttributeModifier(BASE_ATTACK_RANGE_UUID, "bonus", 1, AttributeModifier.Operation.ADDITION));
+        attributes.put(Attributes.ATTACK_KNOCKBACK, new AttributeModifier(BASE_ATTACK_KNOCKBACK_UUID, "bonus", 1, AttributeModifier.Operation.ADDITION));
         return slot == EquipmentSlot.MAINHAND ? attributes.build() : super.getDefaultAttributeModifiers(slot);
     }